Fixed panic on NAME_ERROR
NAME_ERROR responses usually have 0 ANCOUNT and a SOA record in the
Authority section.  The SOA record is usually for '.' (ie, the empty
name).  The appending of +1 in the question processing, instead of
having the empty label +1 the offset, interacted poorly here and caused
the offset to be off by one when trying to process the soa record's
rrtype.